+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.1.0-rc2] - 2024-03-07
+
+### Added
+
+- Release Test with quickstart example
+- CLI Test with quickstart example
 
 ## [0.1.0-rc1] - 2024-01-15